// app/interfaces/IRest.ts
import { Request, Response } from 'express';

export interface IRestController {
  list: (req: Request, res: Response) => void;
  create: (req: Request, res: Response) => void;
  show: (req: Request, res: Response) => void;
  update: (req: Request, res: Response) => void;
  delete: (req: Request, res: Response) => void;
}

export type TParams = {
  search?: string;
  page?: number;
  size?: number;
};

export type TPayload = {
  title: string;
  author: string;
  isbn: string;
  published_year: number;
  genre: string;
  copies_available: number;
  total_copies: number;
  picture: string;
};

export interface IRestModel<T, P = TPayload> {
  list: (params?: TParams) => Promise<T[]>;
  create: (payload: P) => Promise<T>;
  show: (id: string) => Promise<T>;
  update: (id: string, payload: T) => Promise<T>;
  delete: (id: string) => Promise<T>;
}